[#142] feat : 클릭이벤트에 대응되는 스타일 변경에 대해 스토리 생성
