+++ src/@components/Navigation/index.stories.tsx
@@ -1,7 +1,10 @@
 import React from "react";
-import type { ComponentMeta } from "@storybook/react";
+import type { ComponentMeta, ComponentStory } from "@storybook/react";
 
+import Base from "../Core/Base";
 import Navigation from "./index";
+import NavIcon from "./NavIcon";
+import { NAV_ICONS } from "@/constants/navIcons";
 
 export default {
   title: "Navigation/Navigation",
@@ -10,4 +13,19 @@ export default {
 
 export const Default = () => {
   return <Navigation />;
+};
+
+const [firstNavItem] = NAV_ICONS;
+const { activeIcon, defaultIcon, text } = firstNavItem;
+export const OneIcon: ComponentStory<typeof NavIcon> = props => {
+  return (
+    <Base position="relative">
+      <NavIcon {...props} defaultIcon={defaultIcon} activeIcon={activeIcon} />
+    </Base>
+  );
+};
+OneIcon.args = {
+  text,
+  isClicked: false,
+  isLast: false,
 };
